requirements.txt
# Core
langchain>=0.3.0
langchain-community>=0.3.0

# LLM & Embeddings (Local via Ollama)
langchain-ollama>=0.2.0

# Vetorial & RAG
chromadb>=0.5.15
pypdf>=4.2.0

# Interface & Utilidades
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1